backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-telegram-bot==21.6
httpx==0.27.2
python-multipart==0.0.12
pillow==10.4.0
cloudinary==1.41.0
apscheduler==3.10.4
aiofiles==24.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
